0.2.9 update
This is probably going to be the final update for 0.2.x. I have learned a bit more about Python, and so I'm going to rewrite JshBot a third time. The overhaul will probably take place over the summer when I have more time to focus on my own projects, so 0.2.9 is essentially just going to be a small patch to keep things running until then.

* Prevented file writing for every single user upon initialization
* Prevented whole playlists from being downloaded through soundtags